seedmtk is now Pixelv1_sprout
Since second-gen MTK variants differ considerably, unified builds are not possible. Hence, the common codename is obsolete.

The Infinix HOT-2 uses an MTK header in the boot.img, while the Pixelv1 has a standard
AOSP one, and also the processor in the HOT-2 is an MT6580 in place of the Pixelv1's MT6582.

I am trying to get booting builds on the HOT-2, but to no avail as of now.

Device tree is also updated to reflect this change, and sorry for any inconvenience this drastic change has been causing.

Change-Id: I301a7de50b3cae4fd90585c649d6322305ec2e00
Signed-off-by: MSF Jarvis <msfjarvis@gmail.com>